Add pre and post scripts
support for
bump (the whole process)
changelog
commit
tag

These are run if the actual task is also run. The functionality makes
use of a decorator which takes the name of the task as an argument and
uses that to construct the config value (pretag posttag for ex). The
script is passed the new version as an argument. Perhaps in future we
may wish to pass more arguments to the scripts?