Add some coverage exclusions
Exclude some gitops that are basically just git module operations.
Exclude the commandline which uses the components we are testing.
However this might benefit from some integration testing to make sure
the command_line part works when everything is combined
Exclude the printer as it is just a print
Exclude the dunder files
Exclude the custom_cmd_group