Make Versup compatiable with Python 2
Aside from dependency changes, the following has been altered

* `FileNotFoundError` exception replaced with `IOError` throughout
* Strings to print function coerced to unicode with `u` prefix
* Dictionary merge function instead of `{**dict, **dict2}` - to be
  removed once Python2 support is dropped
* Tests have paths and strings coerced into unicode or standard `str`

In the future once Python 2 has become obsolete, these changes can be
undone to simplify the code more.

unfortunately some projects and places still rely heavily on Python 2
and this change is to not exclude them.

One thing to note is that black doesn't work on Python 2. This shouldn't
be a problem as black is only used for formatting.