Check that current branch matches main branch
The config has a new entry under `commit/mainbranch` which is used to
define the main or release branch. Any version update to a branch with a
different name to this config setting will generate a warning.
The user is still able to make the version release, but it will not be
on the configured branch.

FIX #6