No longer write default to home dir
The default config is no longer written entirely to the user's home.
This was a good idea, but overall causes issues with updates to the
config overall. It's much better for the user to explicitly create the
config in their home directory and just use the defaults as a base from
within the application.

The override order of project over global over default still exists.

Adjust documentation and tests accordingly.

FIX #7