Warn about unstaged files earlier
Unstaged files would only prevent a tag being created but the rest of
the versup process would still execute. This patch changes the behaviour
to check for unstaged files at the start of the versup run, prompting
the user if they wish to continue with the unstaged files or not.